github: Run clang test with AddressSanitizer enabled.
This commit is based on a similar one from OVN by Dumitru Ceara:
  a429b24f7bf5 ("ci: Enable AddressSanitizer in Linux clang CI test runs.")

It's useful to run testsuite with address sanitizer enabled to catch
memory leaks and invalid memory accesses.  Skipping re-check if
AddressSanitizer reports are present in the test run directory to
not lose them.

Right now OVS has no memory leaks detected on a testsuite run with -O1.
With -O2 there are few false-positive leak reports in test-ovsdb
application, so not using this optimization level for now.  For the
same reason not enabling leak detection by default for everyone.
Enabled only in CI.

AddressSanitizer increases execution time for this job from ~12 to ~16
minutes, but it looks like a reasonable sacrifice.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>