hindex: remove the next variable in safe loops
Apart from being safer from the point of view of potential undefined
behavior, the multi-variable version of the safe iterators have the
benefit of not requiring the declaration of an external variable to hold
the next position.

This patch removes the NEXT variable from all HINDEX_*_SAFE macros and
all of its callers.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>