ovsdb-idl: Only process successful txn in ovsdb_idl_loop_run.
Otherwise we hide the real transaction result from the user.  This may
cause problems as the user will not detect error cases.  For example,
if the server refuses a transaction with "constraint violation", the
user should be notified because it might need to retry the transaction.

For clients that process database changes incrementally (using change
tracking) that's currently not possible because ovsdb_idl_loop_run()
silently consumes the failure, no other update was received from the
server, so the user will not add anything to the new transaction to
be sent with the next call to ovsdb_idl_loop_commit_and_wait().

Commit 0401cf5f9e06 ("ovsdb idl: Try committing the pending txn in
ovsdb_idl_loop_run.") tried to optimize for the common, successful
case.  Maintain the same approach and optimize for transactions that
succeeded but fall back to the old mechanism of processing failures
within ovsdb_idl_loop_commit_and_wait() instead.

Fixes: 0401cf5f9e06 ("ovsdb idl: Try committing the pending txn in ovsdb_idl_loop_run.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>