ovsdb-idl: Fix use-after-free when destroying an IDL loop.
Transactions that are still incomplete (waiting for a reply from the
server) are kept in the IDL's 'outstanding_txns' map.  When a transaction
is destroyed, ovsdb_idl_txn_destroy() will take care of removing the
transaction from the 'outstanding_txns' map if the transaction was
incomplete.

When destroying the IDL loop, instead of trying to abort the
committing_txn which would set the state to "aborted", just disassemble
it, such that ovsdb_idl_txn_destroy() properly cleans it up.

Fixes: 53a540e5311c ("ovsdb-idl: ovsdb_idl_loop_destroy must also destroy the committing txn.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>