json: Improve json parsing
To parse a json string prior to this change, json_lex_input is called
with each character of the string. If the character needs to be copied
to the buffer, it is copied individually. This is an expensive
operation, as often there are multiple characters in a row
that need to be copied, and copying memory in blocks is more efficient
than byte by byte. To improve this, the string is now copied
in blocks. As long as the parser state stays the same, no copy is
performed. As soon as the state is about to change (for example,
reaching the end of a string), the last n characters are copied to the
buffer.

There is also a conditional that checks if the current character is a
new line, which is quite unlikely. When this was examined with perf, the
comparison had a very high CPU cycle usage. To improve this, the
OVS_UNLIKELY macro was used, which forces the compiler to switch the
order of the instructions.

The json_lex_input function is called for each character of a string,
and calling the function so many times uses a lot of CPU cycles.
Making it an inline function greatly reduced the percentage of cycles
used by the call.

Here is the improvement seen in the json-string-benchmark test:

  SIZE      Q  S         BEFORE       AFTER      CHANGE
--------------------------------------------------------
100000      0  0 :      0.842 ms     0.511 ms   -39.3 %
100000      2  1 :      0.917 ms     0.561 ms   -38.8 %
100000      10 1 :      1.063 ms     0.678 ms   -36.2 %
10000000    0  0 :     85.328 ms    51.998 ms   -39.1 %
10000000    2  1 :     92.555 ms    57.353 ms   -38.0 %
10000000    10 1 :    106.728 ms    69.288 ms   -35.1 %
100000000   0  0 :    955.375 ms   632.300 ms   -33.8 %
100000000   2  1 :   1031.700 ms   689.375 ms   -33.2 %
100000000   10 1 :   1189.300 ms   821.850 ms   -30.0 %

Here Q is probability (%) for a character to be a '\"' and
S is probability (%) to be a special character ( < 32).

Signed-off-by: Rosemarie O'Riorden <roriorden@redhat.com>