ovsdb: Check for ephemeral columns before writing a new schema.
Clustered databases do not support ephemeral columns, but ovsdb-server
checks for them after the conversion result is read from the storage.
It's much easier to recover if this constraint is checked before writing
to the storage instead.

It's not a big problem, because the check is always performed by the
native ovsdb clients before sending a conversion request.  But the
server, in general, should not trust clients to do the right thing.

Check in the update_schema() remains, because we shouldn't blindly
trust the storage.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Reviewed-by: Simon Horman <simon.horman@corigine.com>