ovsdb: Avoid converting database twice on an initiator.
Cluster member, that initiates the schema conversion, converts the
database twice.  First time while verifying the possibility of the
conversion, and the second time after reading conversion request
back from the storage.

Keep the converted database from the first time around and use it
after reading the request back from the storage.  This cuts in half
the conversion CPU cost.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Reviewed-by: Simon Horman <simon.horman@corigine.com>